Add compute_raster_stats command
This computes stats that are used to transform non-uint8 pixel values
to uint8 which is required by most ML backends.

+++ src/rastervision/raster_sources/rasterio_raster_source.py
@@ -40,7 +40,8 @@ class RasterioRasterSource(RasterSource):
         # the partial chip back to full window size.
         partial_chip = load_window(
             self.image_dataset, window.rasterio_format())
-        chip = np.zeros((height, width, partial_chip.shape[2]), dtype=np.uint8)
+        chip = np.zeros((height, width, partial_chip.shape[2]),
+                        dtype=partial_chip.dtype)
         chip[0:partial_chip.shape[0], 0:partial_chip.shape[1], :] = \
             partial_chip
